(Clean) Separate into smaller and more meaningfull files
