+++ src/stochastic_bandit.py
@@ -46,15 +46,15 @@ class BernoulliMAB(MAB):
         return f"BernoulliMAB({self.means})"
 
 
-def RandomBernoulliBandit(Delta,K):
-    """generates a K-armed Bernoulli instance at random where Delta is the gap between the best and second best arm"""
-    maxMean = Delta + np.random.rand()*(1.-Delta)
-    secondmaxMean= maxMean-Delta
-    means = secondmaxMean*np.random.random(K)
-    bestarm = np.random.randint(0,K)
-    secondbestarm = np.random.randint(0,K)
-    while (secondbestarm==bestarm):
-        secondbestarm = np.random.randint(0,K)
-    means[bestarm]=maxMean
-    means[secondbestarm]=secondmaxMean
-    return BernoulliBandit(means)
+# def RandomBernoulliBandit(Delta, K):
+#     """generates a K-armed Bernoulli instance at random where Delta is the gap between the best and second best arm"""
+#     maxMean = Delta + np.random.rand()*(1.-Delta)
+#     secondmaxMean= maxMean-Delta
+#     means = secondmaxMean*np.random.random(K)
+#     bestarm = np.random.randint(0,K)
+#     secondbestarm = np.random.randint(0,K)
+#     while (secondbestarm==bestarm):
+#         secondbestarm = np.random.randint(0,K)
+#     means[bestarm]=maxMean
+#     means[secondbestarm]=secondmaxMean
+#     return BernoulliMAB(means)
